Add Server/Client Logic
This commit adds in the main server loop that essentially just listens for a static bool flag to be flipped. Once done, the server terminates, closing all sockets and exiting all clients

This commit also adds in the socket listener loop. This loop spins up on a new thread and just patiently waits for any new connections. It only allows up to the max specified connections and once thats reached it will not open another socket until a client disconnects.

This commit also also adds in the client logic. The client loop just waits for input, and parses it accordingly: if its not 9 digits or is not a number the loop is broken and the client is disconnected immediately. Otherwise, it caches the number and waits for more.